Implemented dummy modal window
